+++ src/mapStackChips.js
@@ -19,6 +19,7 @@ export const PRESENTED_MAP_STACK_IDS = Object.freeze([
   'bing-labels',
   'esri-imagery',
   'osm',
+  'gibs-daily',
 ]);
 
 /**